#31142 Prime DPM : créer WS pour la liste des agents avec Prime DPM SDJF #31177 Prime DPM : créer WS pour filtrer les services et agents #31140 Prime DPM : WS droits/listeDroitsAgent

+++ src/main/java/nc/noumea/mairie/ptg/dto/VentilPrimeDto.java
@@ -36,7 +36,7 @@ public class VentilPrimeDto extends VentilDto {
 				return String.format("%s", hibObj.getQuantite());
 			case NB_HEURES:
 			case PERIODE_HEURES:
-				return helper.formatMinutesToString(hibObj.getQuantite());
+				return helper.formatMinutesToString(hibObj.getQuantite().intValue());
 		}
 		
 		return null;
